backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
yfinance==0.2.40
beautifulsoup4==4.12.3
pdfplumber==0.11.0
requests==2.32.0
lxml==5.2.0
python-dotenv==1.0.1
supabase==2.7.0
stripe==10.0.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
alembic==1.13.0
apscheduler==3.10.4
